Add missing cast

+++ src/main/java/br/com/etyllica/core/interpolation/Interpolation.java
@@ -3,7 +3,7 @@ package br.com.etyllica.core.interpolation;
 public class Interpolation {
 	
 	public static float factor(long now, long start, long duration) {
-		return (now - start) / duration;
+		return (now - start) / (float)duration;
 	}
 	
 }
